Remove mention of libksane and SANE from description
The user doesn't care about the underlying libraries used, and even if
they did, mentioning them in the description of the app is not an
appropriate place to do this. The description helps users make a
decision about whether they want to use the app or not, and no user ever
chose or rejected Skanlite on the basis of reading in the description
that it uses the libksane and SANE libraries.